change int to big int for the stats

+++ packages/eventindexer/migrations/1666650701_create_stats_table.sql
@@ -2,8 +2,8 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS stats (
     id int NOT NULL PRIMARY KEY AUTO_INCREMENT,
-    average_proof_time int NOT NULL DEFAULT 0,
-    average_proof_reward int NOT NULL DEFAULT 0,
+    average_proof_time bigint NOT NULL DEFAULT 0,
+    average_proof_reward bigint NOT NULL DEFAULT 0,
     num_proofs int NOT NULL default 0,
     num_verified_blocks int NOT NULL default 0,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ,
